@@ -21,21 +21,21 @@
     "@types/draft-js": "^0.10.34",
     "@types/history": "^4.7.2",
     "@types/jest": "^24.0.17",
-    "@types/jsonwebtoken": "^8.3.2",
+    "@types/jsonwebtoken": "^8.3.3",
     "@types/libsodium-wrappers": "^0.7.5",
     "@types/lodash": "^4.14.136",
     "@types/marked": "^0.6.5",
     "@types/modernizr": "^3.5.3",
-    "@types/node": "^12.7.0",
+    "@types/node": "^12.7.1",
     "@types/prismjs": "^1.16.0",
     "@types/qs": "^6.5.3",
     "@types/querystringify": "^2.0.0",
-    "@types/react": "^16.8.25",
+    "@types/react": "^16.9.1",
     "@types/react-dom": "^16.8.5",
     "@types/react-redux": "^7.1.1",
     "@types/react-router": "^5.0.3",
     "@types/react-router-dom": "^4.3.4",
-    "@types/react-stripe-elements": "^1.3.0",
+    "@types/react-stripe-elements": "^1.3.1",
     "@types/stripe-v3": "^3.1.6",
     "@types/turndown": "^5.0.0",
     "@types/yup": "^0.26.22",
@@ -72,8 +72,8 @@
     "prop-types": "^15.6.2",
     "qs": "^6.7.0",
     "querystringify": "^2.1.1",
-    "react": "^15.5.0 || ^16.0.0-rc",
-    "react-dom": "^15.5.0 || ^16.0.0-rc",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-number-format": "^4.2.0",
     "react-redux": "^7.1.0",
     "react-router": "^5.0.0",
@@ -93,7 +93,7 @@
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "husky": "^3.0.2",
+    "husky": "^3.0.3",
     "jest": "^24.8.0",
     "jest-transform-stub": "^2.0.0",
     "parcel-bundler": "^1.12.3",
