@@ -1,24 +1,14 @@
 import { ThemeProvider } from '@material-ui/styles'
 import * as React from 'react'
-import { connect } from 'react-redux'
-import { ApplicationState } from '../../store'
 import { secureMathRandom } from '../../util/secureMathRandom'
 import { theme } from '../theme'
 
-interface PropsFromDispatch {}
-
-interface PropsFromState {
-  ready: boolean
-}
-
-type AllProps = PropsFromDispatch & PropsFromState
-
-class RootFC extends React.Component<AllProps> {
+class Root extends React.Component {
   public render() {
     return <ThemeProvider theme={theme}>{this.props.children}</ThemeProvider>
   }
 
-  public componentWillMount() {
+  public componentDidMount() {
     // Reload the page every ~24h +/- 2h. This is a hack to make sure clients
     // don't end up on an ancient version of the app.
     const sway = 4 * secureMathRandom() * 3600 - 2 * 3600
@@ -26,16 +16,5 @@ class RootFC extends React.Component<AllProps> {
     setTimeout(() => window.location.reload(true), 1000 * reloadAfter)
   }
 }
-
-const mapStateToProps = ({ clientState, keysState }: ApplicationState) => ({
-  ready: clientState.ready
-})
-
-const mapDispatchToProps = {}
-
-const Root = connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(RootFC)
 
 export default Root
