@@ -1,15 +1,14 @@
 import {
+  Button,
   Container,
   createStyles,
   CssBaseline,
   Divider,
   Grid,
-  IconButton,
   makeStyles,
   Paper,
   Theme,
-  Typography,
-  Button
+  Typography
 } from '@material-ui/core'
 import ArrowBack from '@material-ui/icons/ArrowBack'
 import * as React from 'react'
@@ -17,6 +16,7 @@ import { connect } from 'react-redux'
 import * as Router from 'react-router-dom'
 import ClientInit from '../components/ClientInit'
 import MessageBody from '../components/messages/MessageBody'
+import { MessageListItem } from '../components/messages/MessageListItem'
 import { Profile } from '../components/widgets/profile/Profile'
 import { db } from '../db/db'
 import { ApplicationState } from '../store'
@@ -24,7 +24,6 @@ import { messageReadRequest } from '../store/messages/actions'
 import { Balance } from '../store/models/account'
 import { ClientCredentials, ClientProfile } from '../store/models/client'
 import { Message } from '../store/models/messages'
-import { MessageListItem } from '../components/messages/MessageListItem'
 
 interface Props {}
 
@@ -101,41 +100,39 @@ const MessagePageFC: React.FC<AllProps> = ({
 
   return (
     <ClientInit>
-      <React.Fragment>
-        <CssBaseline />
-        <Container className={classes.headerContainer}>
-          <Grid container spacing={1} justify="space-between">
-            <Grid item xs={7}>
-              <Typography variant="h2" component="h2">
-                <strong>
-                  <Router.Link to="/">Umpyre</Router.Link>
-                </strong>
-              </Typography>
-            </Grid>
-            <Grid item xs={5}>
-              <Profile profile={profile} balance={balance} menu />
-            </Grid>
-            <Grid item xs style={{ position: 'relative' }}></Grid>
-            <Grid item xs={12}>
-              <Divider />
-            </Grid>
+      <CssBaseline />
+      <Container className={classes.headerContainer}>
+        <Grid container spacing={1} justify="space-between">
+          <Grid item xs={7}>
+            <Typography variant="h2" component="h2">
+              <strong>
+                <Router.Link to="/">Umpyre</Router.Link>
+              </strong>
+            </Typography>
           </Grid>
-        </Container>
-        <Container>
-          <Button
-            onClick={() => {
-              if (history.length > 0) {
-                history.goBack()
-              } else {
-                history.push('/')
-              }
-            }}
-          >
-            <ArrowBack /> Back
-          </Button>
-          {messageBodies()}
-        </Container>
-      </React.Fragment>
+          <Grid item xs={5}>
+            <Profile profile={profile} balance={balance} menu />
+          </Grid>
+          <Grid item xs style={{ position: 'relative' }}></Grid>
+          <Grid item xs={12}>
+            <Divider />
+          </Grid>
+        </Grid>
+      </Container>
+      <Container>
+        <Button
+          onClick={() => {
+            if (history.length > 0) {
+              history.goBack()
+            } else {
+              history.push('/')
+            }
+          }}
+        >
+          <ArrowBack /> Back
+        </Button>
+        {messageBodies()}
+      </Container>
     </ClientInit>
   )
 }
